fix : id 타입 수정

+++ src/main/java/com/hobak/happinessql/domain/activity/domain/Activity.java
@@ -19,7 +19,7 @@ public class Activity extends BaseTimeEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "activity_id")
-    private int activityId;
+    private Long activityId;
 
     @Column(nullable = false)
     private String name;
@@ -29,7 +29,7 @@ public class Activity extends BaseTimeEntity {
     private Category category;
 
     @Builder
-    public Activity(int activityId, String name, Category category) {
+    public Activity(Long activityId, String name, Category category) {
         this.activityId = activityId;
         this.name = name;
         this.category = category;
